Adds improved pathfinding to tree jumping
Adds improved pathfinding to tree jumping. Unless there isn't an
available tree around, the AI cannot visit a tree it has been to
previously. The jump speed and arc height now have a random range that
can be set in the inspector. The AI can get stuck between jumps on it's
way to it's destination so an exit will be added next commit.